API refaktorálása
