+++ App/_API/models/user.model.js
@@ -16,7 +16,6 @@ User.init(
     jogosultsag: DataTypes.TINYINT,
     reg_token: DataTypes.STRING,
     magassag: DataTypes.INTEGER,
-    suly: DataTypes.DOUBLE,
     cel_suly: DataTypes.DOUBLE,
   },
   { sequelize, tableName: 'felhasznalok', timestamps: false }
