--- App/_API/routes/info.routes.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const infoController = require('../controllers/info.controller');
-
-router.get("/getAllActiveBlog", infoController.getAllActiveBlog);
-router.post("/postBlog", infoController.postBlog);
-
-module.exports = router;
