+++ App/_API/server.js
@@ -9,10 +9,11 @@ app.use(express.json());
 app.use(require('./middleware/access.middleware').auth);
 
 app.use('/auth', require('./routes/auth.routes'));
-app.use('/', require('./routes/dashboard.routes'));
+app.use('/kcal', require('./routes/kcal.routes'));
 app.use('/admin', require('./routes/admin.routes'));
 app.use('/user', require('./routes/user.routes'));
-app.use('/info', require('./routes/info.routes'));
+app.use('/blog', require('./routes/blog.routes'));
+app.use('/weight', require('./routes/weight.routes'));
 
 app.listen(PORT, (err) => {
   if (err) {
